refactor: 기존 redis 테스트 패키지 구조 변경
